Allow Dataptr_or_null and read-only Dataptr in ALTLIST classes
This can be very useful to iterate over an ALTLIST
very quickly. For example, in an ALTLIST class
that implements a sublist, no copying is needed.

This method always returns a const pointer, so
it cannot be used to change the list.

We also forbid writeable Dataptr for ALTREP lists.